fix: 🐛 Better support scaling (16:10 fixed)
This scales the game screen better. I wanted to get height and max-width
set via asteroids.js but couldn't quite seem to get it working and don't
quite have the time right now. PR for the future I hope. In the
meantime, the resolution is set to a fixed 16:10 aspect ratio, which is
good for Macs but not many others. Sorry!